EduTime - Copy/EduTime/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
